requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
idna==3.10
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
telebot==0.0.5
urllib3==2.2.3
